+++ src/engine/utils.rs
@@ -1,8 +1,8 @@
 use glam::Vec2;
 
-pub fn haversine_distance(lat1: f64, lon1: f64, lat2: f64, lon2: f64) -> f64 {
+pub fn haversine_distance(lat1: f32, lon1: f32, lat2: f32, lon2: f32) -> f32 {
     // Earth's radius in meters
-    const EARTH_RADIUS: f64 = 6_371_000.0;
+    const EARTH_RADIUS: f32 = 6_371_000.0;
 
     // Convert degrees to radians.
     let phi1 = lat1.to_radians();
